Added the `Amps and angle encoding.test` `Markdown_1.0.3/` test inherited from lunamark.
The parts of the file concerned with escaping <, >, and preserving
& unescaped inside HTML entities were removed from the test file,
since these parts of the Markdown reader were removed in the slimmed
down version of the library.

Since & is an escaped character in the plain TeX writer, the remainder
of the test was kept intact, as it tests the correct encoding of the &
character inside links.